+++ include/cask/Observable.hpp
@@ -205,6 +205,17 @@ public:
     ObservableRef<T,E> distinctUntilChanged() const;
 
     /**
+     * Given a sequence of emit values suppress duplicate consecutive events
+     * emitted by the source using the provided comparator to check
+     * equality.
+     * 
+     * @param comparator A comparator to use for testing the equality of
+     *                   two values.
+     * @return A new observable which removes duplicate consecutive events.
+     */
+    ObservableRef<T,E> distinctUntilChangedBy(const std::function<bool(const T&, const T&)>& comparator) const;
+
+    /**
      * Transform each element of the stream using the provided transforming predicate
      * function.
      * 
@@ -541,8 +552,13 @@ ObservableRef<T,E> Observable<T,E>::concat(const ObservableRef<T,E>& other) cons
 
 template <class T, class E>
 ObservableRef<T,E> Observable<T,E>::distinctUntilChanged() const {
+    return distinctUntilChangedBy([](auto left, auto right) { return left == right; });
+}
+
+template <class T, class E>
+ObservableRef<T,E> Observable<T,E>::distinctUntilChangedBy(const std::function<bool(const T&, const T&)>& comparator) const {
     auto self = this->shared_from_this();
-    return std::make_shared<observable::DistinctUntilChangedObservable<T,E>>(self);
+    return std::make_shared<observable::DistinctUntilChangedObservable<T,E>>(self, comparator);
 }
 
 template <class T, class E>